+++ src/Spike.cpp
@@ -80,11 +80,14 @@ struct Spike : public biexpand::Expandable<bool> {
     std::array<sp::ClockTracker, NUM_CHANNELS> clockTracker = {};
     std::array<dsp::TTimer<float>, NUM_CHANNELS> nextTimer = {};
     dsp::SchmittTrigger resetTrigger;
+    // XXX array of triggers when polyphonic
     dsp::SchmittTrigger nextTrigger;
+    dsp::SchmittTrigger prevTrigger;
     dsp::PulseGenerator resetPulse;  // ignore clock pulses when reset is high for 1ms
 
     bool connectEnds = false;
     bool keepPeriod = false;
+    bool allowReverseTrigger = false;
 
     /// @brief Address used by Segment2x8 to paint current step
     int activeIndex = 0;
@@ -274,10 +277,13 @@ struct Spike : public biexpand::Expandable<bool> {
         const bool isNextNormalled = isClockConnected && !isNextConnected;
         const int numSteps = readBuffer().size();
         // update our next trigger normalled to the clock if not connected
+        const float nextCv = inputs[INPUT_NEXT].getNormalPolyVoltage(0.F, channel);
         const bool isNextTriggered =
-            isNextConnected ? nextTrigger.process(inputs[INPUT_NEXT].getNormalVoltage(0.F, channel))
-                            : isClockTriggered;
-        if (isNextTriggered) {
+            isNextConnected ? nextTrigger.process(nextCv) : isClockTriggered;
+        const bool isPrevTriggered =
+            isNextConnected && allowReverseTrigger ? prevTrigger.process(-nextCv) : false;
+        const bool isTriggered = isNextTriggered || isPrevTriggered;
+        if (isTriggered) {
             float period =
                 isNextNormalled ? nextTimer[channel].getTime() : clockTracker[channel].getPeriod();
             nextTimer[channel].reset();
@@ -289,10 +295,12 @@ struct Spike : public biexpand::Expandable<bool> {
                 : 0.F;
         const int curStep = stepDetectors[channel].getCurrentStep();
 
-        float phase = fmodf(
-            static_cast<float>(isNextTriggered + curStep + clamp(stepFraction, 0.F, 0.9999F)) /
-                numSteps,
-            1.F);
+        const float phase =
+            fmodf(static_cast<float>(
+                      rack::math::eucMod(curStep + isNextTriggered - isPrevTriggered, numSteps) +
+                      clamp(stepFraction, 0.F, 0.9999F)) /
+                      numSteps,
+                  1.F);
         return phase;
     }
     void process(const ProcessArgs& args) override
@@ -333,6 +341,7 @@ struct Spike : public biexpand::Expandable<bool> {
         json_object_set_new(rootJ, "usePhasor", json_integer(usePhasor));
         json_object_set_new(rootJ, "connectEnds", json_integer(connectEnds));
         json_object_set_new(rootJ, "keepPeriod", json_integer(keepPeriod));
+        json_object_set_new(rootJ, "allowReverseTrigger", json_boolean(allowReverseTrigger));
         json_object_set_new(rootJ, "polyphonyChannels", json_integer(polyphonyChannels));
         return rootJ;
     }
@@ -343,6 +352,8 @@ struct Spike : public biexpand::Expandable<bool> {
         if (usePhasorJ != nullptr) { usePhasor = (json_integer_value(usePhasorJ) != 0); };
         json_t* connectEndsJ = json_object_get(rootJ, "connectEnds");
         if (connectEndsJ != nullptr) { connectEnds = (json_integer_value(connectEndsJ) != 0); };
+        json_t* allowReverseTriggerJ = json_object_get(rootJ, "allowReverseTrigger");
+        if (allowReverseTriggerJ) { allowReverseTrigger = json_is_true(allowReverseTriggerJ); }
         json_t* keepPeriodJ = json_object_get(rootJ, "keepPeriod");
         if (keepPeriodJ != nullptr) { keepPeriod = (json_integer_value(keepPeriodJ) != 0); };
         json_t* polyphonyChannelsJ = json_object_get(rootJ, "polyphonyChannels");
@@ -460,6 +471,8 @@ struct SpikeWidget : public SIMWidget {
         menu->addChild(createBoolPtrMenuItem("Use Phasor as input", "", &module->usePhasor));
         menu->addChild(createBoolPtrMenuItem("Connect Begin and End", "", &module->connectEnds));
 #endif
+        menu->addChild(createBoolPtrMenuItem("Negative 'next' pulse steps in reverse direction", "",
+                                             &module->allowReverseTrigger));
         menu->addChild(
             createBoolPtrMenuItem("Remember speed after Reset", "", &module->keepPeriod));
     }